Training for the first time - way too slow.

+++ dataset.py
@@ -37,7 +37,7 @@ class Vessel12Dataset(torch.utils.data.Dataset):
         resimg = self.img[idx - 4:idx + 5, :, :]
         resmask = self.mask[idx, :, :]
 
-        return torch.from_numpy(resimg).float(), torch.from_numpy(resmask).float()
+        return torch.from_numpy(resimg).float(), torch.from_numpy(resmask).long()
 
     def __len__(self):
         if self.img is None:
